requirements.txt
python-telegram-bot==22.5
aiohttp==3.13.2
python-dotenv==1.0.0
